Initial CppZip checkin with the current state:
CppZip has some functions that may work in some "good" situations. A lot
of work has to be done, that CppZip has:
 - a basic set of functions to work with zip files, 
 - to be stable and 
 - to be correct.

With the step of publishing CppZip to GitHub, the work can begin:
there are a lot of ideas of features in my head, that wants to be
implemented.